+++ apps/nest-app/src/app.module.ts
@@ -2,11 +2,11 @@ import { Module } from '@nestjs/common';
 import { SaleModule } from './sale/sale.module';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AuthModule } from './sale/ui/auth/auth.module';
-import { UsersModule } from './users/users.module';
+import { StaffsModule } from './sale/ui/staffs/staffs.module';
 import { AuthController } from './sale/ui/auth/auth.controller';
 
 @Module({
-  imports: [TypeOrmModule.forRoot(), SaleModule, AuthModule, UsersModule],
+  imports: [TypeOrmModule.forRoot(), SaleModule, AuthModule, StaffsModule],
   controllers: [AuthController],
   providers: [],
 })
