+++ apps/nest-app/src/sale/ui/auth/auth.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AuthService } from './auth.service';
-import { UsersModule } from '../../../users/users.module';
+import { StaffsModule } from '../staffs/staffs.module';
 import { PassportModule } from '@nestjs/passport';
 import { LocalStrategy } from './local.strategy';
 import { JwtModule } from '@nestjs/jwt';
@@ -9,7 +9,7 @@ import { JwtStrategy } from './jwt.strategy';
 
 @Module({
   imports: [
-    UsersModule,
+    StaffsModule,
     PassportModule,
     JwtModule.register({
       secret: jwtConstants.secret,
