+++ apps/nest-app/src/sale/ui/staffs/staffs.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { StaffsService } from './staffs.service';
+
+@Module({
+  providers: [StaffsService],
+  exports: [StaffsService],
+})
+export class StaffsModule {}
