--- apps/nest-app/src/users/users.module.ts
@@ -1,8 +1,0 @@
-import { Module } from '@nestjs/common';
-import { UsersService } from './users.service';
-
-@Module({
-  providers: [UsersService],
-  exports: [UsersService],
-})
-export class UsersModule {}
